requirements.txt
# APT Emulation Framework Dependencies

# Core dependencies
requests>=2.28.0
psutil>=5.9.0
pynput>=1.7.6

# Windows-specific dependencies (install conditionally)
pywin32>=304; sys_platform == "win32"
wmi>=1.5.1; sys_platform == "win32"

# Optional dependencies for advanced features
dnspython>=2.2.0
urllib3>=1.26.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0